// angular-frontend/src/app/clients/open-f1/models/meeting/meeting.properties.ts
export enum MeetingProperties {
    CIRCUIT_KEY = "circuit_key",
    CIRCUIT_SHORT_NAME = "circuit_short_name",
    COUNTRY_CODE = "country_code",
    COUNTRY_KEY = "country_key",
    COUNTRY_NAME = "country_name",
    DATE_START = "date_start",
    GMT_OFFSET = "gmt_offset",
    LOCATION = "location",
    MEETING_CODE = "meeting_code",
    MEETING_KEY = "meeting_key",
    MEETING_NAME = "meeting_name",
    MEETING_OFFICIAL_NAME = "meeting_official_name",
    YEAR = "year",       
}