// angular-frontend/src/app/clients/open-f1/models/session/session.properties.ts
export enum SessionProperties {
    CIRCUIT_KEY = "circuit_key",
    CIRCUIT_SHORT_NAME = "circuit_short_name",
    COUNTRY_CODE = "country_code",
    COUNTRY_KEY = "country_key",
    COUNTRY_NAME = "country_name",
    DATE_END = "date_end",
    DATE_START = "date_start",
    GMT_OFFSET = "gmt_offset",
    LOCATION = "location",
    MEETING_KEY = "meeting_key",
    SESSION_KEY = "session_key",
    SESSION_NAME = "session_name",
    SESSION_TYPE = "session_type",
    YEAR = "year",       
}